add viem to package.json - rename the sc to agentMemoryContract
