+++ auto-kol/agent/package.json
@@ -28,6 +28,7 @@
         "sqlite3": "^5.1.7",
         "twitter-api-v2": "^1.15.1",
         "uuid": "11.0.3",
+        "viem": "^2.21.54",
         "winston": "^3.11.0",
         "zod": "^3.22.4"
     },
