+++ auto-kol/agent/src/utils/agentMemoryContract.ts
